Refactor preparation of initial states. (#40)
