+++ tests/test_simulate.py
@@ -3,7 +3,8 @@ import pandas as pd
 import pytest
 from pandas.api.types import is_categorical_dtype
 from sid.config import INDEX_NAMES
-from sid.simulate import _prepare_params
+from sid.preparation import prepare_initial_states
+from sid.preparation import validate_params
 from sid.simulate import _process_assort_bys
 from sid.simulate import _process_initial_states
 from sid.simulate import get_simulate_func
@@ -26,10 +27,11 @@ def test_simple_run(params, initial_states, tmp_path):
     initial_infections = pd.Series(index=initial_states.index, data=False)
     initial_infections.iloc[0] = True
 
+    states = prepare_initial_states(initial_states, initial_infections, params)
+
     simulate = get_simulate_func(
         params,
-        initial_states,
-        initial_infections,
+        states,
         CONTACT_MODELS,
         path=tmp_path,
     )
@@ -59,14 +61,14 @@ def test_prepare_params(params):
     params = params.copy().append(s)
 
     with pytest.raises(ValueError, match="No NaNs allowed in the params index."):
-        _prepare_params(params)
+        validate_params(params)
 
 
 @pytest.mark.unit
 @pytest.mark.parametrize("input_", [pd.Series(dtype="object"), (), 1, [], {}, set()])
 def test_prepare_params_with_wrong_inputs(input_):
     with pytest.raises(ValueError, match="params must be a DataFrame."):
-        _prepare_params(input_)
+        validate_params(input_)
 
 
 @pytest.mark.unit
@@ -74,7 +76,7 @@ def test_prepare_params_not_three_dimensional_multi_index(params):
     params = params.copy().reset_index(drop=True)
 
     with pytest.raises(ValueError, match="params must have the index levels"):
-        _prepare_params(params)
+        validate_params(params)
 
 
 @pytest.mark.unit
@@ -82,7 +84,7 @@ def test_prepare_params_no_duplicates_in_index(params):
     params = params.copy().append(params)
 
     with pytest.raises(ValueError, match="No duplicates in the params index allowed."):
-        _prepare_params(params)
+        validate_params(params)
 
 
 @pytest.mark.unit
@@ -91,4 +93,4 @@ def test_prepare_params_value_with_nan(params):
     params["value"].iloc[0] = np.nan
 
     with pytest.raises(ValueError, match="The 'value' column of params must not"):
-        _prepare_params(params)
+        validate_params(params)
